Inserindo valores

+++ spring-data-jpa/spring-data/src/main/java/br/com/francisco/spring/data/repository/CargoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CargoRepository extends CrudRepository<Cargo, Long> {
+public interface CargoRepository extends CrudRepository<Cargo, Integer> {
 }
